Track and report wallet transaction clones
Adds a "walletconflicts" array to transaction info; if
a wallet transaction is mutated, the alternate transaction id or ids are reported there (usually the array will be empty).

Metadata from the original transaction is copied to the mutant, so the transaction time and "from" account of the mutant are reported correctly.