build: fix broken boost chrono check on some platforms
If clock_gettime is implemented outside of libc (librt in this case), configure would fail when testing boost. Since clock_gettime is not present on all OSs, boost only uses it when it can. Check for it in librt and add it to LIBS if found, but don't fail if it's not (since boost won't be expecting it in this case).

Also, reverse the link order as necessary for static libs.

Note that it's possible that there are other similar cases for boost, which may be handled the same way.